+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - align existing workflows with workflow framework concepts
 - auto-adjust workflow:state on Flaw.save
+- unrestrict workflow state names (OSIDB-5049)
 
 ### Removed
 - deprecate workflow manipulation endpoints